Changing to proper sentence casing. (#2)
Because it is very easy to change proper sentence casing into upper case, lower case, and first letter of each word capitalization it is optimal to store test strings in true Title case. This is because it is very difficult to format strings into proper title case.

This allows the string to be used in proper title case that can easily be changed to fit the users styling. But then if the styling required is title case it's formatted as such.